+++ src/observe.js
@@ -317,7 +317,7 @@
      */
 
     function getWatchOptions( newOptions ) {
-        var options = !isDefinedObject( newOptions ) ? {} : newOptions;
+        var options = getDefaultObject( newOptions, {} );
 
         options.timeout = getDefaultNumber( options.timeout, 250 );
         options.starts = getDefaultDate( options.starts, null );
@@ -444,6 +444,10 @@
 
     function getDefaultArray( value, defaultValue ) {
         return isDefinedArray( value ) ? value : defaultValue;
+    }
+
+    function getDefaultObject( value, defaultValue ) {
+        return isDefinedObject( value ) ? value : defaultValue;
     }
 
     function getDefaultStringOrArray( value, defaultValue ) {
@@ -749,8 +753,8 @@
      * 
      * @returns     {Object}                                                The Observe.js class instance.
      */
-    this.setConfiguration = function( newOptions ) {
-        _configuration = !isDefinedObject( newOptions ) ? {} : newOptions;
+    this.setConfiguration = function( newConfiguration ) {
+        _configuration = getDefaultObject( newConfiguration, {} );
         
         buildDefaultConfiguration();