+++ docs/CHANGE_LOG.md
@@ -7,8 +7,8 @@
 
 #### **Configuration Options:**
 - Added a new configuration option called "objectErrorText", which states the error text that should be shown when an object error is detected (defaults to "Errors in object: {{error_1}}, {{error_2}}").
-- Added a new configuration option called "attributeNotValidErrorText", which states the error text that should be shown when a binding object is'nt valid (defaults to "The attribute '{{attribute_name}}' is not a valid object.").
-- Added a new configuration option called "attributeNotSetErrorText", which states the error text that should be shown when a binding attribute is'nt set (defaults to "The attribute '{{attribute_name}}' has not been set correctly.").
+- Added a new configuration option called "attributeNotValidErrorText", which states the error text that should be shown when a binding object isn't valid (defaults to "The attribute '{{attribute_name}}' is not a valid object.").
+- Added a new configuration option called "attributeNotSetErrorText", which states the error text that should be shown when a binding attribute isn't set (defaults to "The attribute '{{attribute_name}}' has not been set correctly.").
 
 #### **General Improvements:**
 - All console error logging now goes through the correct method.